Handles if virtualdub (wine stuff) is installed or not.

+++ otrverwaltung3p/path.py
@@ -65,20 +65,27 @@ def get_tools_path(filename=""):
 
 
 def get_internal_virtualdub_path(filename=""):
+    vdub_path = getdatapath("tools/intern-VirtualDub", filename)
     if os.path.expanduser("~") in os.path.abspath(sys.path[0]):
         # started from home dir
-        return getdatapath("tools/intern-VirtualDub", filename)
+        if os.path.exists(vdub_path):
+            return vdub_path
+        else:
+            return  None
     else:
         # started from the system
-        if os.path.exists(os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", 'VERSION')):
-            if not filecmp.cmp(getdatapath("tools/intern-VirtualDub", 'VERSION'),
-                               os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", 'VERSION')):
-                # Version ist nicht aktuell
-                shutil.rmtree(os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub"), ignore_errors=True)
+        if os.path.exists(vdub_path):
+            if os.path.exists(os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", 'VERSION')):
+                if not filecmp.cmp(getdatapath("tools/intern-VirtualDub", 'VERSION'),
+                                   os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", 'VERSION')):
+                    # Version ist nicht aktuell
+                    shutil.rmtree(os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub"), ignore_errors=True)
+                    shutil.copytree(getdatapath('tools/intern-VirtualDub'),
+                             os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub"), symlinks=True)
+            else:
                 shutil.copytree(getdatapath('tools/intern-VirtualDub'),
                          os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub"), symlinks=True)
-        else:
-            shutil.copytree(getdatapath('tools/intern-VirtualDub'),
-                     os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub"), symlinks=True)
 
-        return os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", filename)
+            return os.path.join(BaseDirectory.xdg_data_home, "otrverwaltung3p/intern-VirtualDub", filename)
+        else:
+            return None
